Added dictionary-based admin list filter options
In order to support additional options for admin list filters, a
new list filter format was added. This dictionary-based format includes
the pre-existing options of `field` and `class` as well as a new
`kwargs` option. When incldued, the value for this key will be passed to
the list filter class's `__init__` as keyword arguments.

This format will be used to implement the feature requested in #8851 and
could also be used to solve #16311 as well.